using the agent as worker
